requirements.txt
appdirs==1.4.4
asgiref==3.2.3
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.4.4
awsebcli==3.19.2
backcall==0.2.0
beautifulsoup4==4.9.1
botocore==1.19.47
bs4==0.0.1
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.7
django-admin-thumbnails==0.2.5
django-allauth==0.39.1
django-ckeditor==5.9.0
django-countries==5.3.3
django-crispy-forms==1.11.1
django-currencies==0.10.1
django-debug-toolbar==1.10.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-storages==1.11.1
django-utils-six==2.0
filelock==3.0.12
future==0.16.0
gunicorn==20.1.0
idna==2.6
importlib-metadata==3.10.1
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jmespath==0.10.0
jsonfield==2.1.1
jupyter-client==6.1.7
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.0.1
packaging==20.4
parso==0.7.1
pathspec==0.5.9
pep8==1.7.1
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.8
py==1.9.0
pycodestyle==2.5.0
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.6.1
python-decouple==3.1
python-gettext==4.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.24.0
requests-oauthlib==1.2.0
s3transfer==0.1.13
semantic-version==2.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.0
stripe==1.77.2
termcolor==1.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==20.4.2
wcwidth==0.1.9
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1
